==> notebooks/02_validate_data.ipynb <==
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = 'all'

import pandas as pd
from pathlib import Path
import pyarrow.parquet as pq

month = 1
year = 2024

raw_data_dir = Path('..') / 'data' / 'raw'
raw_data_file_path = raw_data_dir / f'citi_bike_rides_{year}_{month:02}.parquet'

table = pq.read_table(raw_data_file_path)
citi_bike_rides = table.to_pandas()
citi_bike_rides.head()

citi_bike_ride_cp = citi_bike_rides.copy()
citi_bike_ride_cp['duration'] = citi_bike_ride_cp['ended_at'] - citi_bike_ride_cp['started_at']
citi_bike_ride_cp.head()

citi_bike_ride_cp['duration'].describe().T

citi_bike_ride_cp['duration'].quantile(0)
citi_bike_ride_cp['duration'].quantile(0.01)
citi_bike_ride_cp['duration'].quantile(0.995)
citi_bike_ride_cp['duration'].quantile(0.999)

duration_filter = (citi_bike_ride_cp['duration'] > pd.Timedelta(0)) & (citi_bike_ride_cp['duration'] <= pd.Timedelta(hours = 5))
sum(~duration_filter)

sorted_df = citi_bike_ride_cp.sort_values(by = 'started_at', ascending = True)  

# Get the top 10 (smallest) and bottom 10 (largest) values  
top_10 = sorted_df.head(10)  
bottom_10 = sorted_df.tail(10)  

top_10

bottom_10

filter_date_range = (citi_bike_ride_cp['started_at'] >= "2024-01-01") & (citi_bike_ride_cp['started_at'] < "2024-02-01")
sum(~filter_date_range)

final_filter = duration_filter & filter_date_range
numbers_dropped = final_filter.shape[0] - sum(final_filter) # numbers dropped
numbers_dropped
numbers_dropped/final_filter.shape[0] * 100

citi_bike_rides = citi_bike_rides[final_filter]
citi_bike_rides = citi_bike_rides[['started_at', 'start_station_id']]
#citi_bike_rides.rename(columns = {'started_at': 'pickup_datetime',
#                                  'start_station_id': 'pickup_location_id'}, inplace=True)
citi_bike_rides.head()

year = 2024
month = 1
path = Path("..") / "data" / "processed" / f"rides_{year}_{month:02}.parquet"
citi_bike_rides.to_parquet(path, engine="pyarrow", index=False)
