==> notebooks/05_transform_raw_data_to_featur_target.ipynb <==
%load_ext autoreload
%autoreload 2

# Show all output for a cell
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

import sys
import os
import numpy as np
import pandas as pd

sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))

# Now you can import from src
from src.data_utils import load_and_process_citibike_data

citi_bike_rides = load_and_process_citibike_data(year=2024)

citi_bike_rides.head()
citi_bike_rides.tail()

from src.data_utils import transform_raw_data_into_ts_data

citi_bike_ts_data = transform_raw_data_into_ts_data(citi_bike_rides)
citi_bike_ts_data.head()

citi_bike_ts_data.shape

citi_bike_ts_data.head()

citi_bike_ts_data.tail()

from src.data_utils import transform_ts_data_into_features_and_target_loop

features, targets = transform_ts_data_into_features_and_target_loop(citi_bike_ts_data, window_size=24*28, step_size=24)

features

from src.data_utils import transform_ts_data_into_features_and_target_loop

features, targets = transform_ts_data_into_features_and_target_loop(citi_bike_ts_data, window_size=24*28*1, step_size=24)

tabular_data = features
tabular_data["target"] = targets

from src.config import TRANSFORMED_DATA_DIR
tabular_data.to_parquet(TRANSFORMED_DATA_DIR / "citi_bike_tabular_data.parquet", engine="pyarrow")
