==> notebooks/10_load_features_hopsworks.ipynb <==
from dotenv import load_dotenv
import os
load_dotenv()

import sys
import os

# Add the parent directory to the Python path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))

from datetime import datetime
import pandas as pd
from src.data_utils import load_and_process_citibike_data

from_year = 2024
# to_year = datetime.now().year
to_year = 2025
print(f"Download raw data from {from_year} to {to_year}")

rides = pd.DataFrame()
chunks = []
# for year in range(from_year, to_year+1):

#     rides_one_year = load_and_process_citibike_data(year)

#     chunks.append(rides_one_year)
#     #break

for year in range(from_year, to_year+1):
    for month in range(1, 13):
        try:
            rides_one_month = load_and_process_citibike_data(year, months=[month])
            chunks.append(rides_one_month)
        except Exception as e:
            print(f"Skipping {year}-{month:02}: {e}")


# Concatenate all chunks at the end
rides = pd.concat(chunks, ignore_index=True)
print("Data loading complete.")

rides

rides.shape

from src.data_utils import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)

ts_data.info()

import hopsworks

api_key = os.getenv('HOPSWORKS_API_KEY')  
project_name = os.getenv('HOPSWORKS_PROJECT_NAME')  

# pip install confluent-kafka
# Initialize connection to Hopsworks  
project = hopsworks.login(  
    api_key_value=api_key,  
    project=project_name  
)  
print(f"Successfully connected to Hopsworks project: {project_name}")

feature_store = project.get_feature_store()

FEATURE_GROUP_NAME = "citi_bike_time_series_hourly_feature_group"
FEATURE_GROUP_VERSION = 1

feature_group = feature_store.get_or_create_feature_group(
    name=FEATURE_GROUP_NAME,
    version=FEATURE_GROUP_VERSION,
    description="Time series data at hourly freaquency",
    primary_key=["start_station_id", "hour"],
    event_time="hour"
)

feature_group.insert(ts_data, write_options={"wait_for_job": False})
