==> notebooks/12_model_training_pipeline.ipynb <==
%load_ext autoreload
%autoreload 2

# Show all output for a cell
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

import sys
import os

# Add the parent directory to the Python path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))
import src.config as config

import hopsworks

project = hopsworks.login(
    project=config.HOPSWORKS_PROJECT_NAME, api_key_value=config.HOPSWORKS_API_KEY
)

feature_store = project.get_feature_store()
feature_group = feature_store.get_feature_group(
    name=config.FEATURE_GROUP_NAME,
    version=config.FEATURE_GROUP_VERSION
)

# Create a feature view if it doesn't already exist
try:
    feature_store.create_feature_view(
        name=config.FEATURE_VIEW_NAME,
        version=config.FEATURE_VIEW_VERSION,
        query=feature_group.select_all(),
    )
    print(f"Feature view '{config.FEATURE_VIEW_NAME}' (version {config.FEATURE_VIEW_VERSION}) created successfully.")
except Exception as e:
    print(f"Error creating feature view: {e}")

# Retrieve the feature view
try:
    feature_view = feature_store.get_feature_view(
        name=config.FEATURE_VIEW_NAME,
        version=config.FEATURE_VIEW_VERSION,
    )
    print(f"Feature view '{config.FEATURE_VIEW_NAME}' (version {config.FEATURE_VIEW_VERSION}) retrieved successfully.")
except Exception as e:
    print(f"Error retrieving feature view: {e}")

citi_bike_ts_data, _ = feature_view.training_data(description="Time series hourly citi bike rides")

citi_bike_ts_data = citi_bike_ts_data.sort_values(["start_station_id", "hour"]).reset_index(drop=True)

citi_bike_ts_data.head()

citi_bike_ts_data["hour"].min()

citi_bike_ts_data["hour"].max()

citi_bike_ts_data.shape

citi_bike_ts_data['start_station_id'].value_counts()

citi_bike_ts_data[citi_bike_ts_data['start_station_id'] == '5329.03']

import pandas as pd
citi_bike_ts_data["hour"] = pd.to_datetime(citi_bike_ts_data["hour"], errors="coerce")

citi_bike_ts_data["hour"] = citi_bike_ts_data["hour"].dt.tz_localize(None)

citi_bike_ts_data["year_month"] = citi_bike_ts_data["hour"].dt.to_period("M")

gte = citi_bike_ts_data["year_month"] >= pd.Period("2024-01", freq="M")
lte = citi_bike_ts_data["year_month"] <= pd.Period("2024-12", freq="M")
cond = gte & lte
filtered_data = citi_bike_ts_data[cond].reset_index(drop=True)

filtered_data.drop(columns=["year_month"], inplace=True) 

ts_data = filtered_data

ts_data.head()

from src.data_utils import transform_ts_data_into_features_and_target

features, targets = transform_ts_data_into_features_and_target(ts_data, window_size=24*28, step_size=23)

features.sort_values(["start_station_id", "hour"])

features_copy = features.copy()

features_targets = features.copy()
features_targets["target"] = targets

features_targets.shape

from datetime import datetime, timedelta  
import pandas as pd  
from src.data_utils import split_time_series_data  

# Define the cutoff date as 28 days before today  
cutoff_date = pd.Timestamp(datetime.now() - timedelta(days=28))  
# # Convert the cutoff date to a string in a specific format (e.g., 'YYYY-MM-DD')  
# cutoff_date_str = cutoff_date.strftime('%Y-%m-%d')
# # Print the string representation of the cutoff date  
# print(f"Cutoff date as string: {cutoff_date_str}")
cutoff_date

X_train, y_train, X_test, y_test = split_time_series_data(
    features_targets,
    cutoff_date=cutoff_date,
    target_column="target"
)

print(X_train.shape)
print(y_train.shape)
print(X_test.shape)
print(y_test.shape)

from src.pipeline_utils import get_pipeline
pipeline = get_pipeline()

pipeline.fit(features, targets)

from sklearn.metrics import mean_absolute_error
predictions = pipeline.predict(features)

test_mae = mean_absolute_error(targets, predictions)
print(f"{test_mae:.4f}")

import joblib  


# Save the pipeline  
joblib.dump(pipeline, config.MODELS_DIR / "lgb_model.pkl")

from hsml.schema import Schema
from hsml.model_schema import ModelSchema

input_schema = Schema(features)
output_schema = Schema(targets)
model_schema = ModelSchema(input_schema=input_schema, output_schema=output_schema)

model_registry = project.get_model_registry()

model = model_registry.sklearn.create_model(
    name="citi_bike_demand_predictor_next_hour",
    metrics={"test_mae": test_mae},
    description="LightGBM regressor",
    input_example=features.sample(),
    model_schema=model_schema
    #include_files=["src/", "requirements.txt"]  # Include directories and files
)
# https://community.hopsworks.ai/t/attributeerror-windowspath-object-has-no-attribute-startswith/1003
# model.save(config.MODELS_DIR / 'lgb_model.pkl')
model.save('//Users//yashmathur//Documents//MS_DS//Python_Spring_25//CDA_500//Test_Final//models//lgb_model.pkl')

from src.inference import load_model_from_registry

model = load_model_from_registry()

from src.inference import get_model_predictions

preds = get_model_predictions(model, features_copy)
preds

test_mae = mean_absolute_error(targets, preds["predicted_demand"])
print(f"{test_mae:.4f}")
